core/standards/scripts/scheduler2/location_hash/601.html

<!doctype html>
<title>Location hash - set hash from iframe - check result after page completely loaded</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
var histlength;
t.step(function() {
    if( location.hash && location.hash != '#' ) {
        location.href = location.href.replace(/#.*$/,'');
        return;
    }
    histlength = history.length;
});
</script>
<div id="log"></div>
<iframe src="601i.html"></iframe>
<script src="delay.php?delay=1"></script>
<script>
window.onload = function () {
    setTimeout(completely_loaded, 100);
};
function completely_loaded() {
    t.step(function() {
        assert_equals( location.hash, '#foo' );
        assert_equals( history.length, histlength );
    });
    t.done();
}
</script>
